Fix up an incorrect refactoring of Location
I didn't look deeply enough at Location previously to realise that City
already extended it. Given that, this change effectively reverts the
previous refactor of Location to put the changes where they should be,
on City.
