+++ Cmfcmf/OpenWeatherMap/Util/Location.php
@@ -33,14 +33,6 @@ class Location
      */
     public $lon;
 
-    public $name;
-
-    public $localNames = [];
-
-    public $country;
-
-    public $state;
-
     /**
      * Create a new location object.
      *
@@ -49,19 +41,9 @@ class Location
      *
      * @internal
      */
-    public function __construct(
-        $lat = null,
-        $lon = null,
-        string $name = null,
-        array $localNames = [],
-        string $country = '',
-        string $state = ''
-    ) {
+    public function __construct($lat = null, $lon = null)
+    {
         $this->lat = isset($lat) ? (float)$lat : null;
         $this->lon = isset($lon) ? (float)$lon : null;
-        $this->name = $name ?? '';
-        $this->localNames = $localNames ?? [];
-        $this->country = $country ?? '';
-        $this->state = $state;
     }
 }
